+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # main
+- New program:      Accelerometer2GraceL1b.
+- New program:      Orbit2GraceL1b.
+- New program:      SatelliteTracking2GraceL1b.
+- New program:      StarCamera2GraceL1b.
 - New option:       conditionStringContains/MatchPattern: supports wildcards.
 - Bugfix:           ConditionFileExist: fixed bug with wildcards.
 - Bugfix:           GUI: fixed MatchCaseSensitive for schema selction.